Add functionality to fetch torrent content items along with the hierarchy.

+++ Torrenter/include/torrenter/peer_info.h
@@ -17,8 +17,8 @@ struct PeerInfo_
     std::string client;
     int up_rate;
     int down_rate;
-    int total_down;
-    int total_up;
+    long long int total_down;
+    long long int total_up;
     float progress;
     int connection_type;
 };
